♻️ refactor(device): consume shared working directory schema
