--- apps/server/src/routers/lambda/workingDirSchema.ts
@@ -1,1 +1,0 @@
-export { workingDirConfigSchema } from '@lobechat/types';
